Change route for submitLook to /gallery/submitLook. Check if session is started before redirection.  /gallery/submitLook/formLook only with session started
