--- app/gallery/submit/page.jsx
@@ -1,10 +1,0 @@
-
-// Se não tiver sessão iniciada deverá ser redirecionado para o login
-// Exemplo: twovest.com/gallery/submit
-const Submit = () => {
-  return (
-    <div>Submit</div>
-  )
-}
-
-export default Submit
